moved test ydk file renamed test data dir. Renamed test scripts
